requirements.txt
asn1crypto==1.3.0
cryptography==2.8
Django==3.0.3
djangorestframework==3.11.0
enum34==1.1.6
idna==2.7
ipaddress==1.0.23
keyring==21.1.0
keyrings.alt==3.4.0
psycopg2==2.8.4
pycrypto==2.6.1
pygobject==3.26.1
pytz==2019.3
pyxdg==0.26
SecretStorage==3.1.2
six==1.14.0
